pyproject.toml
[project]
name = "master-defence-project"
version = "1.0.0"
description = "基于 RAG 的智能聊天机器人"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "openai>=1.6.0",
    "langchain-community>=0.0.10",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.5.0",
]

[project.scripts]
dev = "uvicorn backend.main:app --reload --host 0.0.0.0 --port 8001"
